+++ plugins/postcss-custom-properties/src/lib/is-ignored.ts
@@ -1,12 +1,12 @@
 function isBlockIgnored(ruleOrDeclaration) {
-	var rule = ruleOrDeclaration.selector ?
+	const rule = ruleOrDeclaration.selector ?
 		ruleOrDeclaration : ruleOrDeclaration.parent;
 
 	return /(!\s*)?postcss-custom-properties:\s*off\b/i.test(rule.toString());
 }
 
 function isRuleIgnored(rule) {
-	var previous = rule.prev();
+	const previous = rule.prev();
 
 	return Boolean(isBlockIgnored(rule) ||
 		previous &&